requirements.txt
aiofiles==0.5.0
alembic==1.5.6
aniso8601==7.0.0
arrow==1.1.0
astroid==2.5.1
async-exit-stack==1.0.1
async-generator==1.10
bcrypt==3.2.0
binaryornot==0.4.4
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cookiecutter==1.7.2
cryptography
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.63.0
google-api-core==1.26.1
google-auth==1.28.0
google-cloud-core==1.6.0
google-cloud-storage==1.36.2
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
httptools==0.1.1
idna
inflect==5.3.0
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
jinja2-time==0.2.0
lazy-object-proxy==1.5.2
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.23
packaging==20.9
passlib==1.7.4
Pillow==8.1.2
poyo==0.5.0
promise==2.3
protobuf==3.15.5
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.1
pylint==2.7.2
PyMySQL==1.0.2
pyparsing==2.4.7
pysodium==0.7.9
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
python-slugify==5.0.0
pytz==2021.1
PyYAML==5.4.1
qrcode==6.1
requests==2.25.1
rsa==4.7.2
Rx==1.6.1
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.13.6
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.7.4.3
ujson==5.1.0
ulid-py==1.1.0
urllib3==1.26.3
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
wrapt==1.12.1
google==3.0.0
